Update README.md about local mobile running
I had problems running my game from a mobile device on the same network,
and the problem was fixed by opening TCP ports 3000 and 3001 on my
system's firewall. This commit notes that these ports should be open if
one wants to run their game from a mobile device on the same network.